# applications/passport-service/swagger/contact-service-swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Basic resource service used for exposing data from database.",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "contact-service"
  },
...
  "schemes" : [ "https" ],
  "paths" : {
    "/v1/users/{userId}/contacts" : {
      "get" : {
        "tags" : [ "contact" ],
        "summary" : "get contacts",
        "description" : "Get a list of contacts",
        "operationId" : "getContacts",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ContactResponse"
              }
            }
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      },
      "post" : {
        "tags" : [ "contact" ],
        "summary" : "add contact",
        "description" : "Create a new contact and return it",
        "operationId" : "createContact",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ContactRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "schema" : {
              "$ref" : "#/definitions/ContactResponse"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "500" : {
            "description" : "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "AddressRequest" : {
      "type" : "object",
      "required" : [ "country", "street", "town" ],
      "properties" : {
        "country" : {
          "type" : "string",
          "example" : "USA"
        },
        "town" : {
          "type" : "string",
          "example" : "New York"
        },
        "street" : {
          "type" : "string",
          "example" : "17th Street"
        },
        "number" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 84
        },
        "entrance" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 3
        },
        "flatNumber" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 123
        }
      },
      "description" : "Address"
    },
    "AddressResponse" : {
      "type" : "object",
      "required" : [ "country", "id", "street", "town" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 31
        },
        "country" : {
          "type" : "string",
          "example" : "USA"
        },
        "town" : {
          "type" : "string",
          "example" : "New York"
        },
        "street" : {
          "type" : "string",
          "example" : "17th Street"
        },
        "number" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 84
        },
        "entrance" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 3
        },
        "flatNumber" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 123
        }
      },
      "description" : "Address"
    },
    "ContactRequest" : {
      "type" : "object",
      "required" : [ "firstName", "lastName" ],
      "properties" : {
        "firstName" : {
          "type" : "string",
          "example" : "Stéphane",
          "description" : "First name"
        },
        "lastName" : {
          "type" : "string",
          "example" : "Maldini",
          "description" : "Last name"
        },
        "email" : {
          "type" : "string",
          "example" : "stéphane@email.com",
          "description" : "Email",
          "pattern" : ".+@.+\\..+"
        },
        "relationship" : {
          "type" : "string",
          "example" : "friend",
          "description" : "Relationship"
        },
        "phoneNumbers" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PhoneNumberDetails"
          }
        },
        "addresses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/AddressRequest"
          }
        }
      },
      "description" : "User Contacts"
    },
    "ContactResponse" : {
      "type" : "object",
      "required" : [ "firstName", "id", "lastName", "userId" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "example" : 3,
          "description" : "Contact id"
        },
        "userId" : {
          "type" : "string",
          "example" : "jlong",
          "description" : "Used id"
        },
        "firstName" : {
          "type" : "string",
          "example" : "Stéphane",
          "description" : "First name"
        },
        "lastName" : {
          "type" : "string",
          "example" : "Maldini",
          "description" : "Last name"
        },
        "email" : {
          "type" : "string",
          "example" : "stéphane@email.com",
          "description" : "Email",
          "pattern" : ".+@.+\\..+"
        },
        "relationship" : {
          "type" : "string",
          "example" : "friend",
          "description" : "Relationship"
        },
        "phoneNumbers" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PhoneNumberDetails"
          }
        },
        "addresses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/AddressResponse"
          }
        }
      },
      "description" : "User Contacts"
    },
    "PhoneNumberDetails" : {
      "type" : "object",
      "required" : [ "number" ],
      "properties" : {
        "number" : {
          "type" : "string",
          "example" : "+10723123456"
        },
        "description" : {
          "type" : "string",
          "example" : "phone number description Stéphane Maldini",
          "minLength" : 1,
          "maxLength" : 2147483647
        }
      },
      "description" : "Phone Number Details"
    }
  }
}